refactor(worktrees): align Effect module conventions
