+++ apps/server/src/orchestration/Layers/ProviderCommandReactor.test.ts
@@ -62,7 +62,7 @@ import * as NodeServices from "@effect/platform-node/NodeServices";
 import * as Clock from "effect/Clock";
 import { ServerSettingsService } from "../../serverSettings.ts";
 import { VcsStatusBroadcaster } from "../../vcs/VcsStatusBroadcaster.ts";
-import { WorktreeLifecycleLock } from "../../vcs/WorktreeLifecycleLock.ts";
+import * as WorktreeLifecycleLock from "../../vcs/WorktreeLifecycleLock.ts";
 import * as GitWorkflowService from "../../git/GitWorkflowService.ts";
 
 const asProjectId = (value: string): ProjectId => ProjectId.make(value);
@@ -272,7 +272,7 @@ describe("ProviderCommandReactor", () => {
       }),
     );
     const worktreeLifecycleEvents: string[] = [];
-    const worktreeLifecycle = WorktreeLifecycleLock.of({
+    const worktreeLifecycle = WorktreeLifecycleLock.WorktreeLifecycleLock.of({
       withLock: (effect) =>
         Effect.sync(() => worktreeLifecycleEvents.push("lock")).pipe(
           Effect.andThen(effect),
@@ -439,7 +439,9 @@ describe("ProviderCommandReactor", () => {
         }),
       ),
       Layer.provideMerge(ServerSettingsService.layerTest()),
-      Layer.provideMerge(Layer.succeed(WorktreeLifecycleLock, worktreeLifecycle)),
+      Layer.provideMerge(
+        Layer.succeed(WorktreeLifecycleLock.WorktreeLifecycleLock, worktreeLifecycle),
+      ),
       Layer.provideMerge(ServerConfig.layerTest(process.cwd(), baseDir)),
       Layer.provideMerge(NodeServices.layer),
     );
